+++ src/Binrep/Example/Tar.hs
@@ -1,7 +1,6 @@
 module Binrep.Example.Tar where
 
 import Binrep
-import Binrep.Util ( natVal'' )
 import Binrep.Generic
 import Binrep.Generic qualified as BR
 import Binrep.Type.Common ( Endianness(..) )
@@ -42,7 +41,7 @@ instance KnownNat n => Put (TarNat n) where
       where
         pfxNulls = B.replicate (fromIntegral pfxNullCount) 0x30
         pfxNullCount = n - blen an - 1
-        n = natVal'' @n
+        n = typeNatToBLen @n
 
 instance KnownNat n => Get (TarNat n) where
     get = do
@@ -51,7 +50,7 @@ instance KnownNat n => Get (TarNat n) where
           0x00 -> return $ TarNat an
           w    -> FP.err $ "TODO expected null byte, got " <> show w
       where
-        n = natVal'' @n
+        n = typeNatToBLen @n
 
 -- Partial header
 data Tar = Tar