+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Rewrite.
   * generics (!!)
   * tests (including for generic derivers)
   * the fastest `Int {- positive -} -> Natural` you ever did see
+  * fumieval's mason for fast, monoidal serializing
+  * use my strongweak library alongside refined for things like numerics
 
 ## 0.1.0 (2022-04-22)
 Initial release.