+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Rewrite.
   * generics (!!)
   * tests (including for generic derivers)
   * the fastest `Int {- positive -} -> Natural` you ever did see
-  * fumieval's mason for fast, monoidal serializing
   * use my strongweak library alongside refined for things like numerics
+  * fumieval's mason for fast, monoidal serializing
+  * András Kovács' flatparse for ultra-fast parsing
 
 ## 0.1.0 (2022-04-22)
 Initial release.